db: upgrade to SQLAlchemy 1.1, fixing invalid PostgreSQL SQL in User.is_default_user
eea19c23b741 triggered a bug in SQLAlchemy 1.0.17, causing invalid SQL to be
emitted for PostgreSQL: For the (Python) query '(x == y) == z', SQLAlchemy
generates the SQL 'x = y = z'. SQLite and MySQL accepts this, but PostgreSQL
does not. This bug has been fixed in 1.1.

Testing and review shows no problems with 1.1