logging: don't change the log level in the handler config
I find it confusing that even though the log level for sqlalchemy.engine was
INFO, because the log level of handler_console_sql was WARN, there were no
INFO output from sqlalchemy.engine.
Therefore remove the level setting for the handlers in the default settings.
I think this makes it (a bit) more intuitive.