style: drop form-horizontal - our style is much closer to plain Bootstrap forms
form-horizontal is made for grid markup. It give form-groups a negative margin
to break out of the grid ... but the way we use it for settings, we have to do
weird things to undo that. The default styling for forms is much closer to what
we want. It looks ok without our custom styling and is easier to style to our
style.

If we want grid markup with form-horizontal, it would be correct to
re-introduce both at once.