contextbar: fix navbar markup
This makes the contextbar markup more consistent with the Bootstrap
documentation at https://getbootstrap.com/docs/3.3/components/#navbar and with
the markup of the main navbar.

This fixes a minor UI issue on small devices where the navbar-collapse border
would show up as a too wide and too low black line.