rcext: this command actually *does* require a config file ... but doesn't require db configuration
'gearbox make-rcext' would fail when trying to do config['here'] without
initializing the configuration.

The command would work if we gave it a whole db session, but it is more correct
to only require the configuration if that is all it need. The command will
however still fail, as the configuration isn't passed on to tg.config .