pygments: reimplement get_lem under the name get_extension_descriptions
The old implementation was cryptic and over-engineered. And reduce() is not a
builtin in Python 3.

This function works on static input, and it was verified that this
implementation returns exactly the same as the old one. And is simpler and
slightly more readable. In my opinion.

Inspired by change by Lars Kruse.