files: simplify header, making it more like changeset
"Search file list" is moved out of the header to a separate div.

The "Follow current branch" functionality is dropped, and there is thus less
need for passing the current URL around ... but it is generally still used for
browser history navigation.

The removal of branch following is a temporary feature regression - that will
be fixed next.