diff: fix crash when displaying diff on a single file
File diff on an image file would fail on
  %if op in 'DM':
because op was None:
  TypeError: 'in <string>' requires string as left operand, not NoneType

But really, if op is None, we also don't want to show invalid "Show images"
links. Thus, guard the whole image display section with having an actual op.

_parse_gitdiff will never return op None, but wrapped_diff is more lazy and
might do that. It could be considered a bug in wrapped_diff, and this change is
just a bad workaround.