less: base .metatag styling on Bootstrap .label styling
Metatags kind of are and kind of look like labels, except some differences:
Labels use smaller but bold font. The sizing is also different, as labels are
not inline-block and will not expand the line height but might overflow into
previous/next line, especially if these also have labels.

We still keep the custom colors for font and background.

Alternatively, we could put the "label" class in the markup and just let the
metatag markup set the colors, similar to what Bootstrap "label-*" does.