less: remove borders and padding around navbar items
Navbar items were styled with borders to look like embossed areas that could be
clicked. The top menu also had it in the drop-down menu, while the context menu
didn't.

The areas are big, and it is quite clear it is some kind of clickable menu ...
and the clickable area is also highlighted when hovering. We can thus do
without the borders and get a slightly cleaner look.