less: make pull request 'next iteration' changeset table smaller and scrollable
This property was lost in 88ce09daea37.

When the whole list is made scrollable, the graph can no longer be positioned
absolute. Instead, make it relative so it scrolls with the content ... and 0
height (with the graph overflowing) so the table content will be shown next to
the overflowing graph.