less: run lessc with --relative-urls so relative paths are rewritten
25e079891ff5 broke select2 - it did for example no longer display the search
symbol in the input field. The css had a relative path to 'select2.png' which
no longer was correct when the literal reference moved to style.css .

Bootstrap had the same issue with '../fonts/glyphicons-halflings-regular.*' ...
but we do not rely much on that.

As a quick fix, just adjust the relative reference and keep pointing at the
original location.

A next step might be to copy files around in a cross platform compliant way,
reliably at the right point in the development/build/install process.