email: fix fallback sending to multiple admins
If no recipients are set, emails are sent to all admins and the address
specified in the config file setting 'email_to'.

However, the existing code did not cater for the fact that multiple
addresses could be set there: the setting was treated as a single string and
passed along as-is.

Adapt the code to split the email_to contents on a comma, as this is the
separator expected by backlash.