tests: issues: use urlify_text wrapper rather than urlify_issues directly
urlify_text is a wrapper around different kinds of urlification methods.
The other tests for urlification use that wrapper, while the new
urlify_issues tests introduced in commit b43e77fa50dd used urlify_issues
directly.

By itself this is not a real problem, but it does not allow to detect the
interplay between different urlification issues. For example, an issue
pattern normally detecting #123 conflicts with the escaping of an apostrophe
' into &#39; . Some new tests exposing this problem are added.