lib: fix detection of &#39; as issue reference
Commit 494c793cc160 changed HTML escaping to please HTML 4 email readers.
The HTML entity '&apos;' was replaced by '&#39;'.
Unfortunately, the pound character '#' is often used to mark issue
references, like 'bug #56'. While this depends on the issue patterns
actually configured, this pattern is so common that we cannot expect users
to set their issue_pat regular expressions such that '&#123;' is not
matched.

Instead, keep the original '&apos;' replacement at first in method html_escape,
but introduce a final step that just replaces '&apos;' with '&#39;'.

The order of replacement in urlify_text then changes from:
    html_escape (to HTML4)
    urlify_issues
to
    html_escape (to HTML5)
    urlify_issues
    make HTML5 more like HTML4

Test coverage show the problem case being solved.